break up things into more components
